requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
requests
python-multipart
structlog
loguru
httpx
async-lru
Pillow
gradio
pytest
pytest-asyncio
httpx
llama-cpp-python
huggingface-hub
